+++ tests/test_flows/raw_sql_scripts/mssql_pytsql_isolate/prep/more_tables.sql
@@ -29,4 +29,5 @@ INNER JOIN (
   FROM {{in_schema}}.raw01 WITH(NOLOCK)
 ) base
   ON apgs.entity = base.entity
-
+CREATE INDEX raw_start_date ON {{out_schema}}.raw01A (start_date DESC)
+CREATE INDEX raw_start_date_end_date ON {{out_schema}}.raw01A (end_date, start_date DESC)